Add files via upload
Attached are screengrabs of the first 37 lines of each worksheet as the full excel file was too large to submit. I also attached a copy of my code in full. At the bottom of the code file includes a list of all outside help received for coding errors and proper syntax.